COMPFEST14 A solved

+++ COMPFEST14/A.cpp
@@ -1,9 +1,9 @@
 #include "bits/stdc++.h"
-
+ 
 using namespace std;
-
+ 
 int main(){
-    int n, m;
+    long long n, m;
     cin >> n >> m;
     cout << (((m-1) == 0) ? ((n-1) * m) : (n * (m-1))) << "\n";
     return 0;
